lotto-52 Created new folder - 'research' in the lotto_app/app/views/ folder and created new files: lotto_app/app/views/research/research_90.py lotto_app/app/views/research/research_8_add.py lotto_app/app/views/research/research.py lotto_app/app/views/research/__init__.py
Now the ResearchViewSet is parent class for the Research90ViewSet and Research8AddViewSet classes.
Created the /{ng}/research_8_add/{pk}/comparison_all_numbers/ endpoint.
